--- .eslintrc.json
@@ -1,31 +1,0 @@
-{
-    "env": {
-      "browser": true,
-      "es6": true,
-      "jest": true
-    },
-    "parser": "@babel/eslint-parser",
-    "parserOptions": {
-      "ecmaFeatures": {
-        "jsx": true
-      },
-      "ecmaVersion": 2018,
-      "sourceType": "module"
-    },
-    "extends": ["airbnb", "plugin:react/recommended", "plugin:react-hooks/recommended"],
-    "plugins": ["react"],
-    "rules": {
-      "react/jsx-filename-extension": ["warn", { "extensions": [".js", ".jsx"] }],
-      "react/react-in-jsx-scope": "off",
-      "import/no-unresolved": "off",
-      "no-shadow": "off"
-    },
-    "ignorePatterns": [
-      "dist/",
-      "build/",
-      "components/",
-      "logic.js/",
-      "setupTest.js",
-      "App.js"
-    ]
-  }
