+++ README.md
@@ -121,7 +121,4 @@ I would like to thank Microverse for thier support
 ## 📝 License <a name="license"></a>
 
 This project is [MIT](https://github.com/Munyabelden/Book-Store/blob/main/LICENSE) licensed.
-
-_NOTE: we recommend using the [MIT license](https://choosealicense.com/licenses/mit/) - you can set it up quickly by [using templates available on GitHub](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository). You can also use [any other license](https://choosealicense.com/licenses/) if you wish._
-
 <p align="right">(<a href="#readme-top">back to top</a>)</p>
