+++ src/components/Book.js
@@ -1,0 +1,11 @@
+const Book = ({ bookList }) => {
+  return (
+    <ul>
+      {bookList.map(book => {
+        return (<li key={book.id}>{book.title} by {book.author}</li>)
+      })};
+    </ul>
+  );
+}
+
+export default Book
