+++ src/components/Books.js
@@ -1,4 +1,5 @@
-import PropTypes from 'prop-types';
+import Form from "./form";
+import Book from "./Book";
 
 const Books = () => {
   const books = [
@@ -27,27 +28,10 @@ const Books = () => {
   return (
     <div className="Books">
       <h2>Books</h2>
-      <ul>
-        {books.map(book => {
-          return (<li key={book.id}>{book.title} by {book.author}</li>)
-        })}
-      </ul>
-      <form>
-        <Input type="text" name="Book Title" id="text" key="Book Title" />
-        <Input type="text" name="Book Author" id="text" key="Book Author" />
-        <button type="submit">Add Book</button>
-      </form>
+      <Book bookList={books}/>
+      <Form />
     </div> 
   );
-}
-
-const Input = ({ name, text}) => {
-  return(<input type={text} name={name} id= {text} placeholder={name} key={name}/>);
-}
-
-Input.prototype = {
-  name: PropTypes.string.isRequired,
-  text: PropTypes.string.isRequired,
 }
 
 export default Books;
